fix: Incorrect file extension for colors test
